# Optimize 2-layer configs for `MultiLayerQG`