@navidcy 2021-11-20
> Looks great to me! It seems possible to reuse one `Params` struct by adding a type parameter `N` corresponding to the number of layers. Then we can define the aliases
> 
> 
> 
> ```julia
> 
> const SingleLayerParams = Params{1}
> 
> const TwoLayerParams = Params{2}
> 
> ```
> 
> 
> 
> I don't see this as necessary for this PR though, just future cleanup.

I agree! Can you open an issue with these ideas for a future cleanup?